Move mockApis directory out of integration_tests
We want the development server, which now depends on these files, to
restart when they are changed. But integration_tests is ignored by
nodemon.
